third_party/WebKit/LayoutTests/fast/forms/fieldset/fieldset-form-collection-radionode-list-expected.txt

CONSOLE WARNING: The <keygen> element is deprecated and will be removed in M57, around March 2017. See https://www.chromestatus.com/features/5716060992962560 for more details.
This test ensures fieldset.elements does NOT contain RadioNodeList.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS owner.elements.length is 24
PASS notRadioNodeList instanceof RadioNodeList is false
PASS successfullyParsed is true

TEST COMPLETE
